Task 6, needs delete fix

+++ src/main/java/by/mishastoma/controller/CrudController.java
@@ -5,7 +5,7 @@ public interface CrudController {
 
     void delete(String obj);
 
-    String findById(int id);
+    String findById(long id);
 
     void update(String obj);
 }
